Move CalendarSelect into its own namespace
