--- catalogue/webapp/components/Calendar/README.md
@@ -1,3 +1,0 @@
-## Purpose
-
-[Edit this on GitHub](https://github.com/wellcometrust/wellcomecollection.org/edit/main/catalogue/components/Calendar/README.md)
