+++ catalogue/webapp/components/CalendarSelect/README.md
@@ -1,0 +1,3 @@
+## Purpose
+
+[Edit this on GitHub](https://github.com/wellcometrust/wellcomecollection.org/edit/main/catalogue/components/CalendarSelect/README.md)
